% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AVGRatio.R
\name{AVGRatio}
\alias{AVGRatio}
\title{AVG Ratio Calculation}
\usage{
AVGRatio(mat.list, Selection_Strength, last.gen = T, num.cores = 1,
  generations = 1:10000 + 20000)
}
\description{
Calculate the AVG ratio for a list of matrices, returning a data.frame
}

