% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AVGRatio.R
\name{AVGRatioSimple}
\alias{AVGRatioSimple}
\title{Simple AVG Ratio Calculation}
\usage{
AVGRatioSimple(mat.list, num.cores = 4)
}
\description{
Calculate the AVG ratio for a list of matrices, returning a data.frame
}

