% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcCorrOmega.R
\name{CalcCorrOmegaEigenVector}
\alias{CalcCorrOmegaEigenVector}
\title{Caclulate Omega Correlation}
\usage{
CalcCorrOmegaEigenVector(mat.list)
}
\description{
Calculates the fitness surface correlation with a list of matrices.
}

