% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalcCorrOmega.R
\name{CalcCorrOmegaKrz}
\alias{CalcCorrOmegaKrz}
\title{Caclulate Omega Correlation}
\usage{
CalcCorrOmegaKrz(mat.list, omega)
}
\description{
Calculates the fitness surface correlation with a list of matrices.
}

