% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalcCorrOmega.R
\name{CalcCorrOmegaRS}
\alias{CalcCorrOmegaRS}
\title{Caclulate Omega Correlation}
\usage{
CalcCorrOmegaRS(mat.list, omega)
}
\description{
Calculates the fitness surface correlation with a list of matrices.
}

