% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalcIsoStat.R
\name{CalcIsoStat}
\alias{CalcIsoStat}
\title{Caculate ISometrics statistic}
\usage{
CalcIsoStat(mat.list, Stat)
}
\description{
Calculates a statistic in the isometric direction given a matrix list and a function
}

