% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalcMeanStat.R
\name{CalcMeanStat}
\alias{CalcMeanStat}
\title{Calculate mean statistic}
\usage{
CalcMeanStat(mat.list, Stat, nsk = 1000)
}
\description{
alculates a statistic in all directions given a matrix list and a function
}

