% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapEffectiveDimension.R
\name{MapEffectiveDimension}
\alias{MapEffectiveDimension}
\title{Effective dimenson wrapper}
\usage{
MapEffectiveDimension(mat.list)
}
\description{
Wrapper to calculate effective dimension in a list
}

