% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NoSelStatMultiPlot.R
\name{NoSelStatMultiPlot}
\alias{NoSelStatMultiPlot}
\title{No directionsal selection plot}
\usage{
NoSelStatMultiPlot(pop.list, StatMap, y.axis, n.traits = 10)
}
\description{
Plots a time series of many populations statistics, averagin by genration
}

