% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NoSelStatMultiPlotMultiPop.R
\name{NoSelStatMultiPlotMultiPop}
\alias{NoSelStatMultiPlotMultiPop}
\title{No directionsal selection plot}
\usage{
NoSelStatMultiPlotMultiPop(drift.list, stab.list, StatMap, y.axis,
  n.traits = 10)
}
\description{
Plots a time series of many populations statistics, averagin by genration and treatment
}

