% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadFolder.R
\name{ReadFolder}
\alias{ReadFolder}
\title{Read Folder}
\usage{
ReadFolder(input.folder, n.traits = 10, sel.type, direct.sel = T)
}
\description{
Read a single population folder and return a list with all poplation atributes
}

