% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadPattern.R
\name{ReadPattern}
\alias{ReadPattern}
\title{Read Pattern from directory}
\usage{
ReadPattern(pattern, n.traits = 10, sel.type, direct.sel = T)
}
\description{
Read many populations from a folder using a pattern
}

