% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadRawPop.R
\name{ReadRawPop}
\alias{ReadRawPop}
\title{Read original data from dat files}
\usage{
ReadRawPop(folder = "/home/diogro/projects/evomod/c-gsl/output/burn_in",
  label = NULL)
}
\description{
Create R objects of the simulated populations for a single generation
}

