% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TimeSeriesMantel.R
\name{TimeSeriesMantel}
\alias{TimeSeriesMantel}
\title{Time series mantel}
\usage{
TimeSeriesMantel(cor.list, num.cores = 4)
}
\description{
Plots sequencial mantel comparisons alog generations
}

