% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AVGRatioPlot.R
\name{AVGRatioPlot}
\alias{AVGRatioPlot}
\title{AVGRatio Plot}
\usage{
AVGRatioPlot(pop.list, modules = FALSE, num.cores = 2,
  x.axis = "selection.strength", x.label = "Peak Movement Rate")
}
\description{
Plots AVGRatio output
}

