% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LastGenStatMultiPlot}
\alias{LastGenStatMultiPlot}
\title{Last Generation Multi Plor}
\usage{
LastGenStatMultiPlot(pop.list, StatMap, y.axis, n.traits = 10)
}
\description{
Plots last generation statistics for one populations in relation to selection strenght
}

