% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LastGenStatMultiPlotWithMean}
\alias{LastGenStatMultiPlotWithMean}
\title{Last Generation Multi Plor with mean}
\usage{
LastGenStatMultiPlotWithMean(pop.list, StatMap, y.axis, n.traits = 10)
}
\description{
Plots last generation statistics for one populations in relation to selection strenght, both directional and mean
}

