% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NoSelStatMultiPlot}
\alias{NoSelStatMultiPlot}
\title{No directionsal selection plot}
\usage{
NoSelStatMultiPlot(pop.list, StatMap, y.axis, n.traits = 10)
}
\description{
Plots a time series of many populations statistics, averagin by genration
}

